requirements.txt
SQLAlchemy == 2.0.0
psycopg2-binary == 2.9.10
werkzeug == 3.1.3
python-dotenv == 1.1.1